Use durable 64-bit Blockwise signing timestamps

+++ prisma/migrations/202609040001_blockwise_signing_timestamp/migration.sql
@@ -1,2 +1,2 @@
-ALTER TABLE "IntegrationOutbox" ADD COLUMN "signingTimestamp" INTEGER;
+ALTER TABLE "IntegrationOutbox" ADD COLUMN "signingTimestamp" BIGINT;
 ALTER TABLE "IntegrationOutbox" ADD COLUMN "signingSignature" TEXT;
